Code refactoring is awesome
I have changed a big part of the source code. Now, we have just to
define vars with their domain and constraints into the state managers... and
that's all. See KQueensManager to see a really simple manager

Obviously, it needs some more tests. I think it is a bit less performant
but the code is so awesome now that I think it's better like that